Move Sound class to a new :model module in Kotlin
