+++ build.gradle
@@ -13,6 +13,7 @@ buildscript {
         classpath 'com.google.gms:google-services:4.3.0'
         classpath 'io.fabric.tools:gradle:1.30.0'
         classpath 'com.google.firebase:perf-plugin:1.3.0'
+        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.41'
 
         classpath 'com.monits:static-code-analysis-plugin:2.6.10'
         classpath 'com.getkeepsafe.dexcount:dexcount-gradle-plugin:0.8.6'
