+++ model/build.gradle
@@ -1,0 +1,10 @@
+apply plugin: 'java-library'
+apply plugin: 'kotlin'
+
+dependencies {
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.3.41'
+    implementation 'androidx.annotation:annotation:1.0.0'
+}
+
+sourceCompatibility = "8"
+targetCompatibility = "8"
